initial implementation (#1)
* initial implementation

* feat: multiplatform support

`ExCurl.Request` now uses the `ExCurl.Zig.MultiPlatform` module
to allow overriding `Zig` opts based on `:os.type()`. This is used
to load `libcurl.dylib` on MacOS instead of the default `libcurl.so`.

* chore: better docs

* chore: add github actions ci

* fix: ci

* fix(ci): install cURL headers

* fix: multiplatform zig now uses load order

The `load_order` module attribute is used as an attempt
to load the first available libcurl from a lsit of known locations
by operating system.