chore: remove Jason usage
Jason was used prior to [0.3.0](https://github.com/danielrudn/ex_curl/pull/13)
to encode the RequestConfiguration struct for Zig.

After 0.3.0, Jason is only used as a test dependency (doctests) and is
included as a peer dependency for zigler.

This commit removed an unused line that still referenced the Jason
module.